Made lerp function representation more sparse to avoid wasting too much space for when elements are not present.
Now it stores only times where it is actually present, but that requieres storing the intervals too.
Access is less direct and needs a bit of computation, so probably a bit slower, but I think the memory saved is worth it, especially in presence-sparse stream graphs.
Maybe I should predict whether dense or sparse storage is better using some kind of metric when parsing.
Maybe I should use the element's presence directly instead of copying it on the weight function.